Task 6 Size

+++ 0x00-hello_world/6-size.c
@@ -8,11 +8,11 @@
 
 int main(void)
 {
-	printf("Size of char: %lu byte(S)\n", sizeof(char));
-	printf("Size of int: %lu byte(S)\n", sizeof(int));
-	printf("Size of long int: %lu byte(S)\n", sizeof(long int));
-	printf("Size of long long int: %lu byte(S)\n", sizeof(long long int));
-	printf("Size of float: %lu byte(S)\n", sizeof(float));
+	printf("Size of char: %lu byte(S)\n", (unsigned long)sizeof(char));
+	printf("Size of int: %lu byte(S)\n", (unsigned long)sizeof(int));
+	printf("Size of long int: %lu byte(S)\n",(unsigned long)sizeof(long int));
+	printf("Size of long long int: %lu byte(S)\n", (unsigned long)sizeof(long long int));
+	printf("Size of float: %lu byte(S)\n",(unsigned long)sizeof(float));
 	return 0
 
 }
